site/requirements.txt
Flask==2.2.2
Pillow~=9.2.0
pytz==2022.2.1
requests==2.28.1
fake-headers
envparse~=0.2.0
